Move the range helpers into ActiveRecord::Filter::RangeHelper
predicate_builder_extension.rb was carrying ~90 lines of range operand
building alongside everything else. That moves to its own file and is mixed
back in, so the methods still live on the predicate builder and the call
sites are unchanged.

  predicate_builder_extension.rb  446 -> 355 lines
  range_helper.rb                        106 lines

Unlike the RangeExtension this PR started with, RangeHelper is a mixin
rather than a utility module called through its own namespace — nothing
reads `RangeHelper.foo(column, value)`, the builder just has the methods.

The include is fully qualified because `module ActiveRecord::Filter::
PredicateBuilderExtension` does not open ActiveRecord::Filter lexically, so
the bare constant would not reach the autoload.
